requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.0
filetype==1.2.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.45.0
google-genai==1.47.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.81
langchain-google-genai==2.1.12
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
starlette==0.49.3
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.39.0
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
